refactor(ast): Introduce a FileParser interface
