+++ src/Mutation/FileMutationGenerator.php
@@ -39,9 +39,9 @@ use Infection\Command\Debug\DumpAstCommand;
 use Infection\FileSystem\FileStore;
 use Infection\Mutator\Mutator;
 use Infection\Mutator\NodeMutationGenerator;
-use Infection\PhpParser\FileParser;
 use Infection\PhpParser\NodeTraverserFactory;
-use Infection\PhpParser\UnparsableFile;
+use Infection\PhpParser\Parser\FileParser;
+use Infection\PhpParser\Parser\UnparsableFile;
 use Infection\PhpParser\Visitor\MutationCollectorVisitor;
 use Infection\Source\Exception\NoSourceFound;
 use Infection\TestFramework\Tracing\Throwable\NoTraceFound;
