+++ src/PhpParser/Parser/PhpParserFileParser.php
@@ -1,0 +1,72 @@
+<?php
+/**
+ * This code is licensed under the BSD 3-Clause License.
+ *
+ * Copyright (c) 2017, Maks Rafalko
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * * Neither the name of the copyright holder nor the names of its
+ *   contributors may be used to endorse or promote products derived from
+ *   this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+declare(strict_types=1);
+
+namespace Infection\PhpParser\Parser;
+
+use Infection\FileSystem\FileStore;
+use PhpParser\Parser;
+use SplFileInfo;
+use Throwable;
+
+/**
+ * @internal
+ */
+final readonly class PhpParserFileParser implements FileParser
+{
+    public function __construct(
+        private Parser $parser,
+        private FileStore $fileStore,
+    ) {
+    }
+
+    public function parse(SplFileInfo $fileInfo): array
+    {
+        try {
+            return [
+                $this->parser->parse(
+                    $this->fileStore->getContents($fileInfo),
+                ) ?? [],
+                $this->parser->getTokens(),
+            ];
+        } catch (Throwable $throwable) {
+            $filePath = $fileInfo->getRealPath() === false
+                ? $fileInfo->getPathname()
+                : $fileInfo->getRealPath()
+            ;
+
+            throw UnparsableFile::fromInvalidFile($filePath, $throwable);
+        }
+    }
+}
